unified Brightness controller to handle laptop/desktop distinction

+++ Source/Utilities/WindowsSettingsBrightnessController.cs
@@ -33,7 +33,7 @@ namespace KLib
             return (int)curBrightness;
         }
 
-        public static void SetBrightness(byte targetBrightness)
+        public static void SetBrightness(int targetBrightness)
         {
             //define scope (namespace)
             System.Management.ManagementScope s = new System.Management.ManagementScope("root\\WMI");
@@ -48,7 +48,7 @@ namespace KLib
 
             foreach (System.Management.ManagementObject o in moc)
             {
-                o.InvokeMethod("WmiSetBrightness", new Object[] { UInt32.MaxValue, targetBrightness }); //note the reversed order - won't work otherwise!
+                o.InvokeMethod("WmiSetBrightness", new Object[] { UInt32.MaxValue, (byte)targetBrightness }); //note the reversed order - won't work otherwise!
                 break; //only work on the first object
             }
